Time: 2 ms (99.76%), Space: 69.4 MB (79.8%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,14 +1,14 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double sum=0;
+        int sum=0;
         for(int i=0;i<k;i++){
             sum+=nums[i];
         }
-        double maxsum=sum;
+        int maxsum=sum;
         for(int i=k;i<nums.length;i++){
             sum+=nums[i]-nums[i-k];
             maxsum=Math.max(sum,maxsum);
         }
-        return maxsum/k;
+        return (double)maxsum/k;
     }
 }
